doubles for d and f

+++ src/pdocks.h
@@ -18,8 +18,8 @@ class PDOCKS {
         byte* finished;
         byte* used;
         double* hittingNumArray;
-        double** D;
-        double** F;
+        float** D;
+        float** F;
         int ALPHABET_SIZE;
         double edgeCount;
         double edgeNum;
@@ -183,11 +183,11 @@ class PDOCKS {
         used = new byte[vertexExp];
         finished = new byte[vertexExp];
         topoSort = new byte[vertexExp];
-        D = new double*[l + 1];
-        double* Dpool = new double[(l+1)* vertexExp];
+        D = new float*[l + 1];
+        float* Dpool = new float[(l+1)* vertexExp];
         for(int i = 0; i < l+1; i++, Dpool += vertexExp) D[i] = Dpool;
-        F = new double*[l + 1];
-        double* Fpool = new double[(l+1)* vertexExp];
+        F = new float*[l + 1];
+        float* Fpool = new float[(l+1)* vertexExp];
         for(int i = 0; i < l+1; i++, Fpool += vertexExp) F[i] = Fpool;
         while (calculatePaths(l, threads)) {
             int imaxHittingNum = calculateHittingNumberParallel(l, false, threads);
